Updated Resources::apply to handle Resource.AllocationInfo.
Previously, `Resource` did not contain `AllocationInfo`. So for
backwards compatibility with old schedulers and tooling, we must
allow operations to contain `Resource`s without an allocation role.
The two interesting cases for adjusting the operation's resource are:

(1) The operation `Resource` does not contain an `AllocationInfo`
    but is being applied to an allocated `Resources`. We allow this
    only if the operation is unambiguous, that is, the allocated
    `Resources` are only allocated to a single role.

(2) The operation `Resource` contains an `AllocationInfo` but is
    being applied to an unallocated `Resources`. In this case we
    simply ignore the `AllocationInfo` of the `Resource`.

Note that we assume no `Resources` store a mix of allocated and
unallocated resources. This is a brittle assumption that we should
have enforcement for.

Review: https://reviews.apache.org/r/55828