Updated resources quantity stripping to strip AllocationInfo.
Currently, `Resources::createStrippedScalarQuantity()` attempts to
create a notion of a "quantity" of resources. In order to do this,
all distinguishing metadata between resources of the same name are
stripped. This currently includes, disk metadata, reservation
metadata (only for dynamic reservations), and shared resource
metadata. To maintain the notion of a quantity, this patch also
strips the allocation metadata.

Review: https://reviews.apache.org/r/55829