Cleanups to the allocator tests.
This was necessary to greatly simplify the changes needed to the
allocator tests as we introduce support for multi-role frameworks.

The main improvement here is to establish and use equality on the
`Allocation` struct, which makes the tests more readable and avoids
the manual probing of the allocation structure across all the tests.

Review: https://reviews.apache.org/r/55868