Adjusted mesos tests to expect empty arrays/maps in protobuf json.
Per MESOS-6568, we now produce empty arrays and maps in json when
there are empty repeated fields and empty maps in protobuf. This
means that some of the tests need to be updated since they either
expect exact json comparisons to match, or expect no additional
keys to be present.

Review: https://reviews.apache.org/r/71907